Moved resolveConflicts to conflicts.service
Added resolveConflicts to peers p2p feature
